image: use custom scheme for docker types
