+++ pkg/image/importer/dockerv1client/client.go
@@ -16,67 +16,15 @@ import (
 	"github.com/docker/distribution/manifest/schema1"
 	"github.com/docker/distribution/manifest/schema2"
 	"github.com/fsouza/go-dockerclient"
-	"k8s.io/klog"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/conversion"
 	knet "k8s.io/apimachinery/pkg/util/net"
 	"k8s.io/client-go/transport"
-	"k8s.io/kubernetes/pkg/api/legacyscheme"
+	"k8s.io/klog"
 
 	"github.com/openshift/library-go/pkg/image/reference"
-	imageapi "github.com/openshift/origin/pkg/image/apis/image"
+
 	"github.com/openshift/origin/pkg/image/apis/image/docker10"
 )
-
-// this is the only entrypoint which deals in github.com/fsouza/go-dockerclient.Image and expects to use our conversion capability to coerce an external
-// type into an api type.  Localize the crazy here.
-// TODO this scheme needs to be configurable or we're going to end up with weird problems.
-func init() {
-	err := legacyscheme.Scheme.AddConversionFuncs(
-		// Convert docker client object to internal object
-		func(in *docker.Image, out *imageapi.DockerImage, s conversion.Scope) error {
-			if err := s.Convert(&in.Config, &out.Config, conversion.AllowDifferentFieldTypeNames); err != nil {
-				return err
-			}
-			if err := s.Convert(&in.ContainerConfig, &out.ContainerConfig, conversion.AllowDifferentFieldTypeNames); err != nil {
-				return err
-			}
-			out.ID = in.ID
-			out.Parent = in.Parent
-			out.Comment = in.Comment
-			out.Created = metav1.NewTime(in.Created)
-			out.Container = in.Container
-			out.DockerVersion = in.DockerVersion
-			out.Author = in.Author
-			out.Architecture = in.Architecture
-			out.Size = in.Size
-			return nil
-		},
-		func(in *imageapi.DockerImage, out *docker.Image, s conversion.Scope) error {
-			if err := s.Convert(&in.Config, &out.Config, conversion.AllowDifferentFieldTypeNames); err != nil {
-				return err
-			}
-			if err := s.Convert(&in.ContainerConfig, &out.ContainerConfig, conversion.AllowDifferentFieldTypeNames); err != nil {
-				return err
-			}
-			out.ID = in.ID
-			out.Parent = in.Parent
-			out.Comment = in.Comment
-			out.Created = in.Created.Time
-			out.Container = in.Container
-			out.DockerVersion = in.DockerVersion
-			out.Author = in.Author
-			out.Architecture = in.Architecture
-			out.Size = in.Size
-			return nil
-		},
-	)
-	if err != nil {
-		// If one of the conversion functions is malformed, detect it immediately.
-		panic(err)
-	}
-}
 
 type Image struct {
 	Image docker.Image
